requirements.txt
streamlit
python-dotenv
openai
PyMuPDF
pytesseract
Pillow
PyPDF2
fpdf
